Conflicts:
	src/FGFDMExec.cpp
	src/FGFDMExec.h
	src/initialization/FGInitialCondition.h
	src/initialization/FGSimplexTrim.cpp
	src/math/FGModelFunctions.h
	src/math/FGNelderMead.cpp
	src/math/FGPropertyValue.cpp
	src/math/FGPropertyValue.h
	src/math/FGQuaternion.h
	src/math/FGStateSpace.cpp
	src/math/FGStateSpace.h
	src/models/FGAccelerations.h
	src/models/FGAerodynamics.h
	src/models/FGAircraft.h
	src/models/FGBuoyantForces.h
	src/models/FGExternalForce.h
	src/models/FGExternalReactions.h
	src/models/FGFCS.cpp
	src/models/FGFCS.h
	src/models/FGFCSChannel.h
	src/models/FGGasCell.h
	src/models/FGGroundReactions.h
	src/models/FGInertial.h
	src/models/FGLGear.h
	src/models/FGMassBalance.h
	src/models/FGOutput.h
	src/models/FGPropagate.cpp
	src/models/FGPropagate.h
	src/models/FGPropulsion.cpp
	src/models/FGPropulsion.h
	src/models/atmosphere/FGWinds.h
	src/models/flight_control/CMakeLists.txt
	src/models/flight_control/FGAccelerometer.h
	src/models/flight_control/FGActuator.h
	src/models/flight_control/FGAngles.cpp
	src/models/flight_control/FGAngles.h
	src/models/flight_control/FGFCSComponent.cpp
	src/models/flight_control/FGFCSComponent.h
	src/models/flight_control/FGFCSFunction.h
	src/models/flight_control/FGFilter.h
	src/models/flight_control/FGGyro.h
	src/models/flight_control/FGMagnetometer.h
	src/models/flight_control/FGPID.h
	src/models/flight_control/FGSensor.h
	src/models/flight_control/FGSensorOrientation.h
	src/models/flight_control/FGSwitch.h
	src/models/flight_control/FGWaypoint.cpp
	src/models/flight_control/FGWaypoint.h
	src/models/flight_control/Makefile.am
	src/models/propulsion/FGElectric.h
	src/models/propulsion/FGEngine.cpp
	src/models/propulsion/FGEngine.h
	src/models/propulsion/FGNozzle.h
	src/models/propulsion/FGPiston.h
	src/models/propulsion/FGPropeller.h
	src/models/propulsion/FGRocket.h
	src/models/propulsion/FGRotor.h
	src/models/propulsion/FGTank.h
	src/models/propulsion/FGThruster.h
	src/models/propulsion/FGTurboProp.h
	src/simgear/props/props.cxx
	src/simgear/props/props.hxx
	src/simgear/xml/easyxml.cxx
	src/utilities/prep_plot.cpp

+++ src/input_output/FGOutputType.cpp
@@ -197,7 +197,7 @@ void FGOutputType::SetRate(double rtHz)
 {
   rtHz = rtHz>1000?1000:(rtHz<0?0:rtHz);
   if (rtHz > 0) {
-    FGModel::SetRate(0.5 + 1.0/(FDMExec->GetDeltaT()*rtHz));
+    FGModel::SetRate(int(0.5 + 1.0/(FDMExec->GetDeltaT()*rtHz)));
     Enable();
   } else {
     FGModel::SetRate(1);
