+++ src/models/FGSurface.cpp
@@ -133,11 +133,11 @@ float FGSurface::GetBumpHeight()
   //height. This is not very fast, but for a beginning.
   //maybe this should be done by interpolating between some precalculated
   //values
-  static const float maxGroundBumpAmplitude=0.4;
-  float h = sin(x)+sin(7*x)+sin(8*x)+sin(13*x);
-  h += sin(2*y)+sin(5*y)+sin(9*y*x)+sin(17*y);
+  static const float maxGroundBumpAmplitude=0.4f;
+  float h = float(sin(x)+sin(7*x)+sin(8*x)+sin(13*x));
+  h += float(sin(2*y)+sin(5*y)+sin(9*y*x)+sin(17*y));
 
-  return h*(1/8.)*bumpiness*maxGroundBumpAmplitude;
+  return float(h*(1./8.)*bumpiness*maxGroundBumpAmplitude);
 }
 
 //%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
